Fix FuzzBuzz target definitions
To build Python with ASAN support as well as instrumentation for fuzzing
is a little bit tricky in this environment.

Yes, you have to disable LeakSanitizer, because the containers don't
seem to be running with ptrace capability, hence
ASAN_OPTIONS=detect_leaks=0.

Also, the environment variables have subtle interaction with `sudo`, and
so it's simplest if we install Python that we build into a user-writable
directory, and dispense with `sudo make install`.

Finally, once Python has been built, to actually build the fuzzing
targets, your easiest way is to use the installed `python3-config`
program to programatically determine the required compilation & linking
options.