Do not build Python with a shared library
Building with --enable-shared ends up requiring us to set
LD_LIBRARY_PATH, or otherwise instruct the loader where to find the
libpython*.so library.  This is a bit of a hassle, and doesn't seem to
get us anything at present with regard to fuzzing, so let's avoid the
shared library.